Replaced chalk with picocolors, replaced prompts with consola. Organized folder structure.
