+++ package.json
@@ -1,53 +1,53 @@
 {
-	"name": "create-bhvr",
-	"version": "0.3.12",
-	"description": "Create a new bhvr project",
-	"type": "module",
-	"main": "./dist/index.js",
-	"types": "./dist/index.d.ts",
-	"bin": {
-		"create-bhvr": "./dist/index.js"
-	},
-	"files": [
-		"dist"
-	],
-	"scripts": {
-		"build": "bun build src/index.ts --outdir dist --target node",
-		"start": "bun ./dist/index.js"
-	},
-	"keywords": [
-		"bun",
-		"vite",
-		"react",
-		"hono",
-		"monorepo",
-		"starter",
-		"template",
-		"create"
-	],
-	"author": "Steve Simkins",
-	"license": "MIT",
-	"devDependencies": {
-		"@types/degit": "^2.8.6",
-		"@types/figlet": "^1.7.0",
-		"@types/fs-extra": "^11.0.4",
-		"@types/inquirer": "^9.0.8",
-		"@types/node": "^20.19.0",
-		"@types/prompts": "^2.4.9",
-		"ts-node": "^10.9.2",
-		"typescript": "^5.8.3"
-	},
-	"dependencies": {
-		"chalk": "^5.4.1",
-		"commander": "^11.1.0",
-		"degit": "^2.8.4",
-		"execa": "^7.2.0",
-		"figlet": "^1.8.1",
-		"fs-extra": "^11.3.0",
-		"ora": "^6.3.1",
-		"prompts": "^2.4.2"
-	},
-	"engines": {
-		"node": ">=18"
-	}
+  "name": "create-bhvr",
+  "version": "0.3.12",
+  "description": "Create a new bhvr project",
+  "type": "module",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "bin": {
+    "create-bhvr": "./dist/index.js"
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "bun build src/index.ts --outdir dist --target node",
+    "start": "bun ./dist/index.js"
+  },
+  "keywords": [
+    "bun",
+    "vite",
+    "react",
+    "hono",
+    "monorepo",
+    "starter",
+    "template",
+    "create"
+  ],
+  "author": "Steve Simkins",
+  "license": "MIT",
+  "devDependencies": {
+    "@types/degit": "^2.8.6",
+    "@types/figlet": "^1.7.0",
+    "@types/fs-extra": "^11.0.4",
+    "@types/inquirer": "^9.0.8",
+    "@types/node": "^20.19.0",
+    "@types/prompts": "^2.4.9",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.3"
+  },
+  "dependencies": {
+    "commander": "^11.1.0",
+    "consola": "^3.4.2",
+    "degit": "^2.8.4",
+    "execa": "^7.2.0",
+    "figlet": "^1.8.1",
+    "fs-extra": "^11.3.0",
+    "ora": "^6.3.1",
+    "picocolors": "^1.1.1"
+  },
+  "engines": {
+    "node": ">=18"
+  }
 }
