+++ src/program.ts
@@ -1,0 +1,3 @@
+import { Command } from "commander";
+
+export const program = new Command();
