+++ src/utils/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_REPO = "stevedylandev/bhvr";
