+++ tsconfig.json
@@ -21,7 +21,11 @@
 		// Some stricter flags (disabled by default)
 		"noUnusedLocals": false,
 		"noUnusedParameters": false,
-		"noPropertyAccessFromIndexSignature": false
+		"noPropertyAccessFromIndexSignature": false,
+    "baseUrl": ".",
+    "paths": {
+      "@/*": ["src/*"]
+    }
 	},
 	"include": ["src/**/*"],
 	"exclude": ["node_modules", "dist", "examples/**/*"]
